CAMLlocal bug in stubs
You cannot use the `CAMLlocal*` macros without an enclosing `CAMLparam*`, otherwise you end up accessing a piece of the stack that was deallocated (and potentially reused). OCaml 4.04.0 guards against this mistake.

Since there is no allocation between the initialization of `ans` and the `return`, you don't need the call to `CAMLlocal1`.